[Refactor] Split menu.spec.ts
